rename react-native components file
